testing loss functions

+++ tests/test_pretraining.py
@@ -33,7 +33,7 @@ print("---------------------------------------------------")
 #-----------Testing loss functions
 text_data= read_txt_file(path= "./data/raw/the-verdict.txt")
 train_data_ratio= 0.90
-split_index= train_data_ratio * len(text_data)
+split_index= int(train_data_ratio * len(text_data))
 train_data_text= text_data[:split_index]
 validation_data_text= text_data[split_index:]
